refactor(infra): BedrockクライアントをcontentProcessing.tsへ分離し、index.tsをバレル化（TDD）
- 新モジュールのエクスポート確認テスト追加
- 後方互換のためindex.tsは再エクスポートのみ

Refs: #9
